fix: codeRabbit feedback

+++ src/main/java/com/example/payment/EmailService.java
@@ -1,5 +1,7 @@
 package com.example.payment;
 
+import java.math.BigDecimal;
+
 public interface EmailService {
-    void sendConfirmation(String email, double amount);
+    void sendConfirmation(String email, BigDecimal amount);
 }
